Release/Back/PT/requirements.txt
> py -m pip install

flask
flask-migrate
flask-sqlalchemy

langchain
langchain-community
youtube_search

python-dotenv

> ollama pull neural-chat